Make Grunt task fail when build fails.
Closes #263.
