+++ test/grunt_tasks/config.js
@@ -1,0 +1,1 @@
+steal.config({});
