+++ test/grunt_tasks/main.js
@@ -1,0 +1,3 @@
+steal(function(){
+	return "hello world!";
+});
