+++ test/grunt_tasks/main_with_error.js
@@ -1,0 +1,3 @@
+bar.baz();
+
+steal(function(){});
